FEAT: Show info of each item on HomeScreen.
Implemented these features.
1. Show brand name, item name, image, and the price of the items.
2. Setup viewmodels properly (path, structure, etc)

This will be the first and the last commit of this branch.

+++ app/src/main/java/com/example/stylish/data/remote/dto/products/Previous.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Previous(
     @SerialName("value")
-    val value: Int? = null,
+    val value: Float? = null,
     @SerialName("text")
     val text: String
 )
